# specs/001-iot-gas-detection/contracts/api.yaml
openapi: 3.0.3
info:
  title: Sistema IoT de Detección de Fugas de Gas API
  description: |
    API REST para monitoreo y control del sistema de detección de fugas de gas.
    Implementa RBAC con 4 roles: admin, operator, technician, auditor.
  version: 1.0.0
  contact:
    name: GEUS Development Team
servers:
  - url: https://{domain}/api/v1
    description: Production server
    variables:
      domain:
        default: api.geus.example.com
security:
  - bearerAuth: []
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List users
      description: Admin only
      parameters:
        - name: role
          in: query
          schema:
            $ref: '#/components/schemas/UserRole'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'

    post:
      tags: [Users]
      summary: Create user
      description: Admin only
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{user_id}:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Users]
      summary: Get user details
      description: Admin only (or self)
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

    patch:
      tags: [Users]
      summary: Update user
      description: Admin only
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

    delete:
      tags: [Users]
      summary: Delete user
      description: Admin only. Soft delete - marks as deleted.
      responses:
        '204':
          description: User deleted
  /users/{user_id}/sensors:
    parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Users]
      summary: Get user's assigned sensors
      description: Admin only (or self for operators)
      responses:
        '200':
          description: Assigned sensors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SensorListResponse'

    put:
      tags: [Users]
      summary: Assign sensors to operator
      description: Admin only. Replaces all assignments.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sensor_ids]
              properties:
                sensor_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        '200':
          description: Sensors assigned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SensorListResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
...
    SensorStatus:
      type: string
      enum: [online, offline, maintenance]
...
    UserRole:
      type: string
      enum: [admin, operator, technician, auditor]
    Sensor:
      type: object
      properties:
        id:
          type: string
          format: uuid
        device_id:
          type: string
        location:
          type: string
        status:
          $ref: '#/components/schemas/SensorStatus'
        wifi_signal:
          type: integer
        mqtt_connected:
          type: boolean
        uptime_seconds:
          type: integer
        test_mode:
          type: boolean
        test_mode_expires_at:
          type: string
          format: date-time
          nullable: true
        correction_factor:
          type: number
        last_reading_at:
          type: string
          format: date-time
          nullable: true
        valve:
          $ref: '#/components/schemas/ValveState'
        dissipator:
          $ref: '#/components/schemas/DissipatorState'
        created_at:
          type: string
          format: date-time
    ValveState:
      type: object
      properties:
        state:
          type: string
          enum: [open, closed]
        mechanical_status:
          type: string
          enum: [ok, stuck, unknown]
        last_state_change:
          type: string
          format: date-time
    DissipatorState:
      type: object
      properties:
        state:
          type: string
          enum: [on, off]
        activation_mode:
          type: string
          enum: [manual, automatic]
        locked_by_alert:
          type: boolean
        last_state_change:
          type: string
          format: date-time
...
    SensorListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Sensor'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          type: string
          enum: [active, suspended]
        notifications_enabled:
          type: boolean
        last_login_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    UserCreateRequest:
      type: object
      required: [email, password, name, role]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
          maxLength: 255
        role:
          $ref: '#/components/schemas/UserRole'
    UserUpdateRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 255
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          type: string
          enum: [active, suspended]
        notifications_enabled:
          type: boolean
    UserListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/User'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...